refactor(pages): add BasePage with shared open/reload, extend MenuPage and CartPage
