+++ src/pages/BasePage.js
@@ -1,0 +1,38 @@
+import { step } from 'allure-js-commons';
+
+export class BasePage {
+  _url;
+  constructor(page) {
+    this.page = page;
+  }
+
+  url() {
+    if (this._url) {
+      return this._url;
+    } else {
+      throw Error(`The property '_url' must be implemented`);
+    }
+  }
+
+  async open() {
+    await step(`Open ${this._pageName()} page`, async () => {
+      await this.page.goto(this.url());
+    });
+  }
+
+  async reload() {
+    await step(`Reload the Menu Page`, async () => {
+      await this.page.reload();
+    });
+  }
+
+  _pageName() {
+    return this.constructor.name.replace('Page', '');
+  }
+
+  async waitForLoading() {
+    await step(`Wait for Cart page to open`, async () => {
+      await this.page.waitForURL('/cart');
+    });
+  }
+}
