+++ src/pages/CartPage.js
@@ -1,8 +1,11 @@
+import { BasePage } from './BasePage';
+
 const { expect, step } = require('@playwright/test');
 
-export class CartPage {
+export class CartPage extends BasePage{
   constructor(page) {
-    this.page = page;
+    super(page);
+    this._url = '/cart';
     this.cartListLocator = page.getByRole('list').nth(1);
     this.notCoffeeMessage = page.getByText('No coffee, go add some.');
     this.totalCheckout = page.getByTestId('checkout');
@@ -34,24 +37,6 @@ export class CartPage {
 
   coffeeListItemAddOneButton(name) {
     return this.page.getByLabel(`Add one ${name}`).nth(1);
-  }
-
-  async open() {
-    await step(`Open the Cart Page`, async () => {
-      await this.page.goto('https://coffee-cart.app/cart');
-    });
-  }
-
-  async waitForLoading() {
-    await step(`Wait for Cart page to open`, async () => {
-      await this.page.waitForURL('/cart');
-    });
-  }
-
-  async reload() {
-    await step(`Reload the Cart Page`, async () => {
-      await this.page.reload();
-    });
   }
 
   async clickCoffeeListItemRemoveAllButton(name) {
