+++ src/pages/MenuPage.js
@@ -1,8 +1,11 @@
+import { BasePage } from './BasePage';
+
 const { expect, step } = require('@playwright/test');
 
-export class MenuPage {
+export class MenuPage extends BasePage{
   constructor(page) {
-    this.page = page;
+    super(page);
+    this._url = '/';
     this.cartLink = page.getByLabel('Cart page');
     this.totalCheckout = page.getByTestId('checkout');
     this.promoMessage = page.getByText(
@@ -21,18 +24,6 @@ export class MenuPage {
   coffeeCupCostLocator(coffeeName) {
     const coffeeCup = this.coffeeCupLocator(coffeeName);
     return this.page.getByRole('listitem').filter({ has: coffeeCup });
-  }
-
-  async open() {
-    await step('Open the Menu Page', async () => {
-      await this.page.goto('/');
-    });
-  }
-
-  async reload() {
-    await step(`Reload the Menu Page`, async () => {
-      await this.page.reload();
-    });
   }
 
   async clickCoffeeCup(coffeeName) {
